--- a/app/__init__.py
+++ b/app/__init__.py
@@ -1,0 +1,1 @@
+# Empty file to make the directory a Python package