--- a/modern-stack/src/components/navigation-header.tsx
+++ b/modern-stack/src/components/navigation-header.tsx
@@ -23,7 +23,6 @@ import {
   User, 
   BarChart3,
   MessageSquare,
-  UserPlus,
   Mail,
   Menu
 } from 'lucide-react';