… app/data/prompts.json
… app/main.py
… app/static/custom.css
… app/static/mobile.css
… auth/api.py
… modern-stack/.eslintignore
… modern-stack/LESSONS_LEARNED.md
… modern-stack/MATRIX_ROADMAP.md
… modern-stack/cookies.txt
… modern-stack/csrf.txt
… modern-stack/prisma/schema.prisma
… modern-stack/scripts/create-admin.js
… modern-stack/src/app/admin/page.tsx
… modern-stack/src/app/api/test-configs/route.ts
… modern-stack/src/app/api/test-encryption/route.ts
… modern-stack/src/app/auth/signin/page.tsx
… modern-stack/src/app/community/page.tsx
… modern-stack/src/app/dashboard/page.tsx
… modern-stack/src/app/invites/page.tsx
… modern-stack/src/app/matrix/page.tsx
… modern-stack/src/app/users/[id]/page.tsx
… modern-stack/src/app/users/create/page.tsx
… modern-stack/src/app/users/invitations/page.tsx
… modern-stack/src/app/users/page.tsx
… modern-stack/src/components/ui/password-strength.tsx
… modern-stack/src/components/ui/progress-indicator.tsx
… modern-stack/src/components/ui/search-help.tsx
… modern-stack/src/components/ui/skeleton.tsx
… modern-stack/src/components/ui/user-credentials.tsx
… modern-stack/src/lib/advanced-search.ts