## Summary

This PR completes the Matrix service refactoring initiative and significantly improves the user creation flow with better authentication, faster responses, and enhanced functionality.

### 🏗️ Matrix Service Refactoring

- **Modular Architecture**: Broke down the monolithic 1,960-line MatrixService into 5 specialized services following Single Responsibility Principle:
  - `MatrixClientService`: Authentication & client management
  - `MatrixMessagingService`: Direct messages & communication  
  - `MatrixEncryptionService`: E2EE management
  - `MatrixSignalBridgeService`: Signal integration
  - `MatrixRoomService`: Room operations

- **Backward Compatibility**: Maintained full API compatibility through coordinating service pattern
- **Environment Config**: Added `parseRooms()` and `parseCategories()` methods for env var parsing
- **Legacy Backup**: Preserved original service as `matrix-legacy.ts`

### 🚀 User Creation & Authentication Improvements

- **Fixed Authentication Issues**: Implemented automatic password reset after user creation to resolve common Authentik auth problems
- **Non-blocking Operations**: Made password reset asynchronous for faster user creation response (67s → ~4s)
- **Matrix Configuration Fix**: Fixed Matrix config loading to check environment variables directly instead of uninitialized service
- **Discourse Integration**: Fixed Discourse service activation and forum post creation during registration
- **Enhanced Timeouts**: Increased Authentik API timeout from 15s to 30s for large user bases

### 🔧 Bug Fixes & Technical Improvements

- **Matrix Config Loading**: Fixed `getConfig` tRPC endpoint returning null despite valid environment variables
- **Service Methods**: Added missing `parseRooms`/`parseCategories` methods after refactoring
- **Lazy Initialization**: Implemented to prevent server startup blocking
- **Debug Logging**: Added comprehensive logging for email and Matrix message templates
- **Error Handling**: Enhanced fallback behavior and error recovery

### 📊 Test Results

- ✅ Matrix service properly shows as configured
- ✅ Discourse forum posts created successfully with URLs
- ✅ User creation now ~94% faster (67s → 4s response time)
- ✅ Password reset working in background
- ✅ All existing Matrix functionality preserved
- ✅ Email templates include Discourse URLs
- ✅ Server startup no longer blocks on Matrix initialization

🤖 Generated with [Claude Code](https://claude.ai/code)